+++ tests/Service/Md5SumTest.php
@@ -5,12 +5,17 @@ namespace MF\PreBuild\Service;
 use MF\PreBuild\AbstractTestCase;
 use MF\PreBuild\Entity\Config;
 use MF\PreBuild\Entity\Md5SumConfig;
+use MF\PreBuild\Entity\Md5SumReplaceConfig;
+use MF\PreBuild\Entity\ReplacePlaceholder;
 
 /**
  * @group unit
  */
 class Md5SumTest extends AbstractTestCase
 {
+    private const MD5_CSS = '6158e1ebc3616886d69778164fc39574';
+    private const MD5_JS = '8d3a465fe1c15cafc1b6c9e2edac5c6a';
+
     /** @var Md5Sum */
     private $md5sum;
 
@@ -29,9 +34,31 @@ class Md5SumTest extends AbstractTestCase
             ])
         );
 
-        $variables = $this->md5sum->findMd5Sum($config);
+        $variables = $this->md5sum->findMd5SumToVariables($config);
 
-        $this->assertSame('6158e1ebc3616886d69778164fc39574', $variables->get('CSSVER'));
-        $this->assertSame('8d3a465fe1c15cafc1b6c9e2edac5c6a', $variables->get('JSVER'));
+        $this->assertSame(self::MD5_CSS, $variables->get('CSSVER'));
+        $this->assertSame(self::MD5_JS, $variables->get('JSVER'));
+    }
+
+    public function testShouldFindMd5SumOfGivenFilesForReplacingPlaceholders(): void
+    {
+        $parametersFile = __DIR__ . '/../Fixtures/Replace/parameters.yml';
+
+        $config = new Config(
+            null,
+            null,
+            new Md5SumReplaceConfig([
+                __DIR__ . '/../Fixtures/Md5/style.css' => [$parametersFile, '___CSSVER___'],
+                __DIR__ . '/../Fixtures/Md5/index.js' => [$parametersFile, '___JSVER___'],
+            ])
+        );
+        $expected = [
+            new ReplacePlaceholder($parametersFile, '___CSSVER___', self::MD5_CSS),
+            new ReplacePlaceholder($parametersFile, '___JSVER___', self::MD5_JS),
+        ];
+
+        $replacements = $this->md5sum->findMd5SumForReplace($config);
+
+        $this->assertEquals($expected, $replacements->toArray());
     }
 }