+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There should always be "Unreleased" section at the beginning. -->
 
 ## Unreleased
+- Add support for parsing md5 sum into file, replacing specific placeholder
 
 ## 2.0.1 - 2020-02-24
 - Fix version in composer json file
@@ -12,10 +13,10 @@
 - Allow Symfony 5
 
 ## 1.1.0 - 2018-02-26
-- Added support for parsing md5 sum
+- Add support for parsing md5 sum
 
 ## 1.0.1 - 2018-01-23
-- Fixed exporting variables
+- Fix exporting variables
 
 ## 1.0.0 - 2018-01-22
 - Initial app